requirements.txt
mysql-connector-python>=8.0.0
ttkbootstrap>=1.0.0